Refactor download functionality into a native plugin for improved performance and maintainability
