+++ android/gradle.properties
@@ -5,3 +5,4 @@ android.enableJetifier=true
 android.builtInKotlin=false
 # This newDsl flag was added automatically by Flutter migrator
 android.newDsl=false
+kotlin.incremental=false
