+++ packages/native_download/pubspec.yaml
@@ -1,0 +1,20 @@
+name: native_download
+description: Native Android downloader for Obtainium.
+publish_to: none
+version: 0.1.0
+
+environment:
+  sdk: ^3.12.0
+  flutter: '>=3.44.0'
+
+dependencies:
+  flutter:
+    sdk: flutter
+
+flutter:
+  plugin:
+    platforms:
+      android:
+        package: dev.imranr.obtainium.native_download
+        pluginClass: NativeDownloadPlugin
+        androidX: true
